+++ plugins.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "lib/bundler_skills"
-require_relative "lib/bundler_skills/command"
-
-BundlerSkills::Hook.register